Move date extension into his own file.
